% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{extractBestSolutionEvaluations.analysis.results}
\alias{extractBestSolutionEvaluations.analysis.results}
\title{TODO}
\usage{
\method{extractBestSolutionEvaluations}{analysis.results}(results, problem,
  search, ...)
}
\description{
TODO
}

