% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{extractBestSolutionEvaluations}
\alias{extractBestSolutionEvaluations}
\title{Extract best solution evaluations}
\usage{
extractBestSolutionEvaluations(results, problem, search, ...)
}
\description{
TODO
}

