% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{getBestSolutionValues}
\alias{getBestSolutionValues}
\title{Get values of best found solutions}
\usage{
getBestSolutionValues(data, problem, search)
}
\arguments{
\item{data}{data object containing the analysis results}

\item{problem}{name of the analyzed problem}

\item{search}{name of the applied search}
}
\value{
Numeric vector containing the values of the best found solutions
  during each run
}
\description{
Get the values of the best found solutions during all runs of the given
\code{search} applied to the given \code{problem}. This is a generic S3
method.
}

