% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{getNumRuns}
\alias{getNumRuns}
\title{Get number of applied search runs}
\usage{
getNumRuns(data, problem, search)
}
\arguments{
\item{data}{data object containing the analysis results}

\item{problem}{name of the analyzed problem}

\item{search}{name of the applied search}
}
\value{
numeric: number of applied search runs
}
\description{
Get the number of applied runs of the given \code{search} when solving the
given \code{problem}. This is a generic S3 method.
}

