% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isMinimizing}
\alias{isMinimizing}
\title{Check if values are being maximized or minimized for a given problem.}
\usage{
isMinimizing(data, problem)
}
\arguments{
\item{data}{analysis results of class \code{james}}

\item{problem}{name of problem}
}
\description{
Compares the first and last obtained value. If contradicting results
are found for different searches or runs solving the same problem, an
error is thrown. Argument 'problem' can be omitted if 'data' contains
results for a single problem only.
}

