\name{james.analysis-package}
\alias{james.analysis-package}
\alias{james.analysis}
\docType{package}
\title{
JAMES Analysis R Package
}
\description{
This package can be used to further analyze and visualize results produced by the analysis tools from the extensions module of the JAMES framework. Functions are provided to plot convergence curves, draw box plots of solution quality or convergence times and to summarize, manipulate or extract data from the results.
}
\details{
\tabular{ll}{
Package: \tab james.analysis\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2015-03-04\cr
License: \tab MIT\cr
}
This version is written for use in combination with version 0.2.1 of the JAMES extensions module.
}
\author{
Herman De Beukelaer <Herman.DeBeukelaer@UGent.be>
}
\references{
JAMES Website: \url{http://www.jamesframework.org}
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
%\examples{
%~~ simple examples of the most important functions ~~
%}
