% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{openPDF}
\alias{openPDF}
\title{Open PDF file}
\usage{
openPDF(file)
}
\arguments{
\item{file}{character: relative path to the PDF file to be opened}
}
\description{
Open a PDF file in the standard PDF viewer.
}
\details{
Opens the given PDF \code{file} in the default PDF viewer. Under Windows,
this is done by calling \code{shell.exec} and whatever program associated
with the PDF file extension will be used. On Unix (including Mac OS X) the
function will use the program named in the option "pdfviewer" (see
\code{help(options)} for information on how this option is set).
}
\examples{
openPDF("pdf/test.pdf")
}

