% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotConvergence}
\alias{plotConvergence}
\title{Plot convergence curves}
\usage{
plotConvergence(data, problem)
}
\arguments{
\item{data}{data object containing the analysis results}

\item{problem}{name of the problem for which the plot is made. Can be omitted
  if the \code{data} contains results for a single problem only.}
}
\description{
Creates a plot showing the convergence curve of each search that has been
applied to the given \code{problem}, averaged over all search runs. If the
\code{data} contains results for a single problem only, the argument
\code{problem} can be omitted.
}

