% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{readAnalysisResults}
\alias{readAnalysisResults}
\title{Read analysis results from JSON file}
\usage{
readAnalysisResults(file)
}
\arguments{
\item{file}{character vector: relative path to a JSON file containing results produced by
the analysis tools in the JAMES extensions module}
}
\value{
S3 object of type "analysis.results" containing the results of running a number of
        searches on a set of problems, performing a number of repeated runs per search
}
\description{
Read results from a JSON file produced by the analysis tools in the JAMES extensions module.
}
\examples{
# get path to raw JSON file included in package distribution
json.file <- system.file("extdata", "results.json", package = "james.analysis")

# read results from file
results <- read.analysis.results(json.file)
}

