% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{readJAMES}
\alias{readJAMES}
\title{Read analysis results from JSON file}
\usage{
readJAMES(file)
}
\arguments{
\item{file}{string: path to a JSON file containing results produced by the
analysis tools from the JAMES extensions module.}
}
\value{
S3 object of class "james" containing the results of running a number
  of searches on a set of problems, where each search has been repeatedly
  applied for a number of runs. Data can be manipulated and extracted using
  the provided functions (see below).
}
\description{
Read results from a JSON file produced by the analysis tools in the JAMES
extensions module.
}
\examples{
# get path to raw JSON file included in package distribution
json.file <- system.file("extdata", "james.json", package = "james.analysis")

# read results from file
james <- readJAMES(json.file)
summary(james)

# plot convergence curves for maize accession data set
plotConvergence(james, problem = "maize-accession", min.time = 100, max.time = 100000)

# create box plots of solution values (quality) and convergence times
boxplot(james, problem = "maize-accession")
boxplot(james, problem = "maize-accession", type = "time")

# extract solution values and convergence times for parallel tempering and random descent
v.pt <- getBestSolutionValues(james, problem = "maize-accession", search = "Parallel Tempering")
t.pt <- getConvergenceTimes(james, problem = "maize-accession", search = "Parallel Tempering")
v.rd <- getBestSolutionValues(james, problem = "maize-accession", search = "Random Descent")
t.rd <- getConvergenceTimes(james, problem = "maize-accession", search = "Random Descent")

# perform wilcoxon test to compare distributions across algorithms
wilcox.test(v.pt, v.rd)
wilcox.test(t.pt, t.rd)
}
\seealso{
Example data: \code{\link{james}}.

  Data access and manipulations methods: \code{\link{reduceJAMES}},
  \code{\link{mergeJAMES}}, \code{\link{getProblems}},
  \code{\link{getSearches}}, \code{\link{getSearchRuns}},
  \code{\link{getNumSearchRuns}}, \code{\link{getBestSolutionValues}},
  \code{\link{getBestSolutions}}, \code{\link{getConvergenceTimes}}.

  Plot functions: \code{\link{plotConvergence}}, \code{\link{boxplot.james}}.
}

