% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{summary.james}
\alias{summary.james}
\title{Print analysis results summary.}
\usage{
\method{summary}{james}(object, d = 3, ...)
}
\arguments{
\item{object}{analysis results of class \code{james}}

\item{d}{number of significant digits (defaults to 3)}
}
\description{
Displays a table with one row per combination of analyzed dataset
and applied search algorithm. The table reports the average (median)
final solution quality across repeated executions and corresponding
standard deviation (inter-quartile range).
}

