% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{getSearches}
\alias{getSearches}
\title{Get names of applied searches}
\usage{
getSearches(data, problem, filter)
}
\arguments{
\item{data}{data object containing the analysis results}

\item{problem}{name of the analyzed problem}

\item{filter}{\link{regular expression} (optional). Only search names that
  match the given regex are returned, if any.}
}
\value{
Vector of strings containing the names of all searches that have been
  applied to the given problem and match the applied filter (if any).
}
\description{
Extracts the names of all searches that have been applied to the given
\code{problem}. If a \code{filter} is set, only those search names matching
the given \link{regular expression} are returned (pattern matching is done
with \code{\link{grep}}). This is a generic S3 method.
}

