\name{james.analysis-package}
\alias{james.analysis-package}
\alias{james.analysis}
\docType{package}
\title{
JAMES Analysis R Package
}
\description{
Analyze and visualize results of studies performed in JAMES, using the analysis tools from the extensions module. Compare algorithm performance, assess influence of parameter values, plot convergence curves, etc.
}
\details{
\tabular{ll}{
Package: \tab james.analysis\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-02-19\cr
License: \tab MIT\cr
}
This version is written for use in combination with version 0.2.1 of the JAMES extensions module.
}
\author{
Herman De Beukelaer <Herman.DeBeukelaer@UGent.be>
}
\seealso{
JAMES Website: \url{http://www.jamesframework.org}
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
%\references{
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
