% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{james}
\alias{james}
\title{Results of example algorithm comparison}
\format{S3 object of class "james", as if produced by
  \code{\link{readJAMES}}.}
\source{
\url{http://www.jamesframework.org/examples/#analysis}
}
\usage{
james
}
\description{
Contains results of an example analysis performed with the JAMES extensions
module. The performance of two algorithms is compared (random descent and
parallel tempering) for a core selection problem in which the mean
entry-to-nearest-entry distance is maximized. Four different data sets have
been analyzed. Details about the performed analysis are provided at the
website (see below).
}
\examples{
# load data
data(james)
summary(james)

# plot convergence curves for maize accession data set
plotConvergence(james, problem = "maize-accession", min.time = 100, max.time = 100000)

# create box plots of solution values (quality) and convergence times
boxplot(james, problem = "maize-accession")
boxplot(james, problem = "maize-accession", type = "time")

# extract solution values and convergence times for parallel tempering and random descent
v.pt <- getBestSolutionValues(james, problem = "maize-accession", search = "Parallel Tempering")
t.pt <- getConvergenceTimes(james, problem = "maize-accession", search = "Parallel Tempering")
v.rd <- getBestSolutionValues(james, problem = "maize-accession", search = "Random Descent")
t.rd <- getConvergenceTimes(james, problem = "maize-accession", search = "Random Descent")

# perform wilcoxon test to compare distributions across algorithms
wilcox.test(v.pt, v.rd)
wilcox.test(t.pt, t.rd)
}
\seealso{
\code{\link{readJAMES}}
}
\keyword{datasets}

