% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{mergeJAMES}
\alias{mergeJAMES}
\title{Merge analysis results}
\usage{
mergeJAMES(data1, data2)
}
\arguments{
\item{data1}{results from the first analysis}

\item{data2}{results from the second analysis}
}
\value{
merged data
}
\description{
Merge results from different analyses. If runs of the same search
applied to the same problem are found in both data sets, these runs
are merged into a single list.
}

