% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotConvergence}
\alias{plotConvergence}
\title{Plot convergence curves}
\usage{
plotConvergence(data, problem, type = c("mean", "median"), col = "black",
  plot.type = "s", lty = NULL, title = "Convergence curve(s)",
  xlab = "Runtime (ms)", ylab = "Value", min.time = NA, max.time = NA,
  legend = TRUE, legend.pos = NULL, legend.inset = c(0.02, 0.05),
  legend.names = NULL, ...)
}
\arguments{
\item{data}{data object containing the analysis results}

\item{problem}{name of the problem for which the plot is made. Can be omitted
if the \code{data} contains results for a single problem only.}

\item{type}{one of "mean" (default) or "median". Determines how the values
from the different search runs are aggregated.}

\item{col}{color(s) of the plotted lines and/or symbols, used cyclically when
providing a vector. Defaults to black.}

\item{plot.type}{defaults to "s" (staircase). See \link{matplot} and
\link{plot} for more information about the possible plot types.}

\item{lty}{line type(s), used cyclically when providing a vector. If set to
\code{NULL}, line types default to 1:n where n is the number of plotted
curves.}

\item{title}{plot title.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{min.time}{zoom in on the part of the curve(s) above this time on the
\code{x-}axis. Defaults to NA in which case no minimum time is set.}

\item{max.time}{zoom in on the part of the curve(s) below this time on the
\code{x-}axis. Defaults to NA in which case no maximum time is set.}

\item{legend}{logical: indicates whether a legend should be added to the
plot. Defaults to TRUE.}

\item{legend.pos}{position of the legend, specified as a keyword  keyword
from the list \code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"},
\code{"left"}, \code{"topleft"}, \code{"top"}, \code{"topright"},
\code{"right"} and \code{"center"}. If set to \code{NULL}, defaults to
\code{"bottomright"} in case values are being maximized or
\code{"topright"} in case of minimization.}

\item{legend.inset}{inset distance(s) from the margins as a fraction of the
plot region. If a single value is given, it is used for both margins; if
two values are given, the first is used for \code{x-}distance, the second
for \code{y-}distance.}

\item{legend.names}{names to be shown in the legend. If set to \code{NULL},
the names default to the search names obtained from calling
\code{\link{getSearches}} on the given \code{data} and \code{problem}.}

\item{...}{optional other arguments passed to \code{\link{matplot}}.}
}
\description{
Creates a plot showing the convergence curve of each search that has been
applied to the given \code{problem}, aggregated over all search runs (mean or
median). This is a generic S3 method.
}
\details{
If the \code{data} contains results for a single problem only, the argument
\code{problem} can be omitted. If desired to plot convergence curves for a
selection of the applied searches, use \link{reduceJAMES} to extract the
respective data.

The curves are plotted using \link{matplot}. More information about the
graphical parameters are provided in the documentation of this function. By
default, a legend is added to the plot. This can be omitted by setting
\code{legend = FALSE}. If desired, a custom legend may then be added. It is
possible to zoom in on a specific region of the plot using the parameters
\code{min.time} and \code{max.time}.

Any additional parameters are passed to \link{matplot}.
}

