% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{readJAMES}
\alias{readJAMES}
\title{Read analysis results from JSON file}
\usage{
readJAMES(file)
}
\arguments{
\item{file}{string: path to a JSON file containing results produced
by the analysis tools from the JAMES extensions module.}
}
\value{
S3 object of class "james" containing the results of running a number
  of searches on a set of problems, where each search has been repeatedly
  applied for a number of runs.
}
\description{
Read results from a JSON file produced by the analysis tools from the JAMES
extensions module.
}
\examples{
# get path to raw JSON file included in package distribution
json.file <- system.file("extdata", "james.json", package = "james.analysis")

# read results from file
james <- readJAMES(json.file)

summary(james)
}

