% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{reduceJAMES}
\alias{reduceJAMES}
\title{Reduce analysis results to selected problems and searches}
\usage{
reduceJAMES(data, problems = ".*", searches = ".*")
}
\arguments{
\item{data}{data object containing the analysis results}

\item{problems}{\link{regular expression} used to filter the problems}

\item{searches}{\link{regular expression} used to filter the searches}
}
\value{
Reduced data set containing only those problems and searches whose
  names match the respective regular expressions.
}
\description{
Reduce the given \code{data} by filtering the analyzed problems and applied
searches based on the given \link{regular expression} (pattern matching is
done with \code{\link{grep}}).
}

